refactor: 예외 클래스명 변경
