+++ src/main/java/step2/dto/WinningCondition.java
