+++ src/main/java/step2/dto/WinningInfo.java
