+++ src/main/java/step2/exception/NotValidLottoNumberException.java
@@ -1,9 +1,9 @@
 package step2.exception;
 
-public class NotValidLottoNumber extends IllegalArgumentException {
+public class NotValidLottoNumberException extends IllegalArgumentException {
     private static final String NOT_VALID_NUMBER = "로또 번호는 1 ~ 45 숫자만 가능합니다.";
 
-    public NotValidLottoNumber() {
+    public NotValidLottoNumberException() {
         super(NOT_VALID_NUMBER);
     }
 }
